Check assert statements with nonsense expressions
Fixes lp:1405778.

I've come across numerous accidental uses of assert where the test
expression is a tuple of the actual test expression and the message,
which always succeeds as the bool of a non empty tuple is always true.

This patch catches the most obvious constants boolean expressions assed
along, certainly not all of them, but there shouldn't be any false
positives.