Update project metadata
Point people to GitHub, not Launchpad. All that's left on Launchpad are
a bunch of stale bugs.

Update the author field to reflect the large number of contributors.
Remove the maintainer field because a bug in something (setuptools?)
overwrites the author field with the maintainer field when the package
gets uploaded to PyPI.