Don't crash with an AttributeError on Windows
Depending on the C library, some constants (like signal.SIGINT) might
not be defined. The previous code attempted to avoid failures on invalid
signals by catching the ValueError raised by signal.signal when given an
invalid signal to handle, however failed to account for the possiblity
of the signal not being defined at all.