Suppress RedefinedWhileUnused for submodule import
Fixes lp:1578051

aec68a7 added module names to error messages, which included
a new class SubmoduleImportation to handle the special case
of submodule imports.  It correctly handled the case of
a submodule import occurring after the root module was imported,
but didnt handle the opposite case of the submodule import
occurring before the root module was imported.