test_errors_syntax: Expect SystemExit error code 1
The test was written to expect SystemExit error code of True.
The SystemExit raised by api.main is of type bool.

True and 1 are equal, and assertEqual treats them that way,
so there is no difference between them.

The other tests in this class expect an int error code,
due to the use of Popen which will convert the SystemExit
bool to an int.  Hence this patch switches to using 1.

More importantly, TestMain.runPyflakes is updated to check
the api.main error code is a bool, and then force the
conversion to int, so that any errors from self.assertEqual
do not show differences of 1 vs True, when in fact assertEqual
considers them to be identical.  These differences of 1 vs True
are distracting to whatever other differences are presented,
especially in error offset bugs like regularly occur, such as
https://github.com/PyCQA/pyflakes/issues/346.