Add check for use of 'is' operator with str, bytes, and int literals
While it often appears to "work", it is considered bad practice to
compare against string, bytes, and int literals using the `is` operator.
This is because it relies on a CPython implementation detail that may
not hold true for other interpreters. For example:

    # Bad
    if x is 'foo':
        ...

    # Good
    if x == 'foo':
        ...

This can easily be caught during static analysis to promote best
practices. This is discussed in the Python bug:
https://bugs.python.org/issue34850