Include column numbers with all error messages (#426)
* Include column numbers with all error messages

As long as the node passed to the message has a col_offset, it is included in
the message. There is already a precedent of printing the column number for
syntax errors.

This changes the API slightly, in that if col_offset is not present on the
node, the message.col is now -1 instead of 0. Otherwise, there would be no way
to distinguish between errors that have no column number and errors with a
column number 0. If this doesn't seem like an acceptable change, I can revert
it, and either always use 0 as the default, or use a separate attribute to
indicate if the message.col is the "real" column number.

I haven't yet tested everything, but I'm not even sure if there are any nodes
that don't have a col_offset. There don't appear to be any represented in the
test suite.

* Simplify some code

* Make column numbers start at 1

This also reverts an API break. message.col now defaults to 0 again, instead
of -1 (0 means there is no column number and so it isn't included in the
printed error message).

* Revert column number API change

It broke a lot of tests.

Now, to minimize API changes as much as possible:

- The message.col attribute starts at 0, same as the AST col_offset
- message.col is also 0 even if col_offset doesn't exist. Note that currently
  there aren't any nodes represented in the pyflakes test suite that don't
  have col_offset. Such nodes do exist in the ast module, but it seems
  unlikely that they would ever result in a message (only expr and stmt nodes
  have it, so for instance a Module node would not).
- The str version of the message starts columns at 1, which is consistent with
  the existing messages for syntax errors, as well as other tools such as
  flake8.

* Fix flake8 issue

Fixes #158 
Fixes #425