Fix bug in async function scope checking
Previously attempting to check an async function scoped tree would
immediately error because tuples are not callable.

The test added here is a direct copy of 'test_scope_function' as
I'm not sure what (if anything) should specifically be tested for
async functions, though it serves to prove the fix.